Remove more COM/WinRT APIs from netstandard ref
